feat: cleaning up py directory
