+++ py/legacy/legacy-3-18-25/README.md
@@ -1,0 +1,3 @@
+# Legacy 3-18-25
+
+The files in this directory were moved from the root directory becuase it was determined that they are no longer used. If you require a file from this directory, please move it back to root.
